Add Gitignore pattern for Cargo and Add Lox interpreter with Rust
Added new gitignore rules for acknowledging compiled files and executables generated by the Cargo package manager.
Also added a new implementation for the Lox interpreter programming language, now in Rust. It is based on the book Crafting Interpreters by Bob Nystrom. The README provides an overview of Lox and explains why Rust is a suitable language for crafting interpreters. The contribution guidelines and license for the project are provided as well.
The notes directory contain a list of topics including inner classes and local classes in Java, lambda expressions, and "effective final" parameters in Java. Also, look into the differences between the keyword 'this' among languages, the use of keyword 'super', the challenges of lexical grammars in Python and Haskell.